Started working on the next section of the Controls Panel.
The section will allow the setting of angle sequences for the robot
arm to move through. Included the buttons that will be used, and encountered
the following issue:

As is, the parent-child three of the app is
                App.tsx
               /       \
    ThreeContainer      ModelControls
                        /            \
                    ValueSlider      AnimatePanel

To manipulate the robot arm, array values are passed from
ValueSlider -> App (via prop callback) -> ThreeContainer (via App setstate)

There are currently inconsistencies where the method 'componentwillupdate'
in ThreeContainer gets called multiple times with a single App setstate.

The deprecated method 'UNSAFE_componentwillreceiveprops' is being used to update values
for the three js model. Although the methods are called twice, since there are
checks based on the previous and new angle values, there are no issues.

The react team suggests to use some static methods that with the current use case,
would require the props in ThreeContainer to be loaded to state. Additionally,
the static methods do not allow the use of class methods, thus 'componentwillupdate'
would be required for updates before render. With this in mind, there would be two
sources of data, on props and state.

As a result, the ThreeContainer.jsx class will be refactored in order to allow for
ModelControls to be a direct child. The needed angle values will be added to the state.
With this, angle updates (via the slider components) will use a callback and directly
update the state of the for the ThreeContainer class. There then only be one source
of data to render, without the need for the static methods or 'UNSAFE_componentwillreceiveprops'.