Three Container Update
Refactored the Three container class by moving all of the model construction methods,
into 'ModelGeneratorFunctions.ts'. Moved all of the important this handles
into an object to allow the construction functions to add models to the three js scene.

Moved the previous state located in App.tsx to 'ThreeContainer' to allow the control panel
and sliders to directly modify state. This was initially done in order to
have on source of data, however that was not the case.

The method 'componentwillupdate' gets called at least twice everytime setstate gets called.
Unfortunately, on both function calls, both the nextState and currectState, respectively are
the same as the previous call. That is:
Call 1 componentwillupdate:  nextState: {x: 4}, prevState: {x: 5}
Call 2 componentwillupdate:  nextState: {x: 4}, prevState: {x:5}

The next state does not change, and due to the fact that the data if statements check for changes,
the functions in each case are run twice.

This was resolved by creating a 'prevState' member object to the ThreeContainer class.
With the object, the previous values are stored and used for the componentwillupdate if statements.
As a result the functions only run once, even with componentwillupdate runs twice.